+++ app/Filament/Admin/Resources/CleaningTasks/RelationManagers/CompletionsRelationManager.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Admin\Resources\CleaningTasks\RelationManagers;
 
+use App\Filament\Concerns\TranslatesRelationManagerTitle;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
@@ -13,6 +14,8 @@ use Filament\Tables\Table;
  */
 class CompletionsRelationManager extends RelationManager
 {
+    use TranslatesRelationManagerTitle;
+
     protected static string $relationship = 'completions';
 
     protected static ?string $title = 'Completion log';