+++ app/Filament/Admin/Resources/Events/RelationManagers/AddOnDayPassesRelationManager.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Admin\Resources\Events\RelationManagers;
 
+use App\Filament\Concerns\TranslatesRelationManagerTitle;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
@@ -14,6 +15,8 @@ use Filament\Tables\Table;
  */
 class AddOnDayPassesRelationManager extends RelationManager
 {
+    use TranslatesRelationManagerTitle;
+
     protected static string $relationship = 'addOnDayPasses';
 
     protected static ?string $title = 'Add-on day passes';