+++ app/Filament/Admin/Resources/Members/RelationManagers/BehaviorNotesRelationManager.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Admin\Resources\Members\RelationManagers;
 
+use App\Filament\Concerns\TranslatesRelationManagerTitle;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
@@ -14,6 +15,8 @@ use Filament\Tables\Table;
  */
 class BehaviorNotesRelationManager extends RelationManager
 {
+    use TranslatesRelationManagerTitle;
+
     protected static string $relationship = 'behaviorNotes';
 
     protected static ?string $title = 'Behavior notes';