+++ app/Filament/Admin/Resources/Members/RelationManagers/MemberUsernameChangesRelationManager.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Admin\Resources\Members\RelationManagers;
 
+use App\Filament\Concerns\TranslatesRelationManagerTitle;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
@@ -12,6 +13,8 @@ use Filament\Tables\Table;
  */
 class MemberUsernameChangesRelationManager extends RelationManager
 {
+    use TranslatesRelationManagerTitle;
+
     protected static string $relationship = 'usernameChanges';
 
     protected static ?string $title = 'Username change log';